added the graph and fixed errors
Added a plot that graphed radii in units of solar radii versus mass in units of solar mass. the graph demonstrated a lot of issues with the code, I had written the equation for gamma X incorecctly I changed the linspace to a logspace so the points were better distributed.